# Increase conv performance by making fewer intermediate arrays