`conv` pads its inputs to use fourier convolution with linear convolution.
Previously, this padding was done by making separate arrays for each pad, and
then concatenating them with the input arrays. I have eliminated the
intermediate pad arrays, and thereby eliminated the creation of two arrays for
each call of `conv`.

I have additionally added a new function, `_circ_conv`, which performs circular
convolution using fourier methods. By separating this out form the `conv`
function, dispatch on the element type can be done my Julia, instead of by a
branch in `conv`. This also allows future use of `_circ_conv` for intentional
circular convolution.