@galenlynch 2019-03-17
This PR results in a sizeable performance increase in `conv`, at the cost of some code readability.

Some simple testing shows that with large input vectors (like I often deal with), the changes increase performance by about 50%, mostly a result of fewer GC interruptions. In addition, memory usage is halved.

Before changes:
```julia
using DSP, BenchmarkTools
julia> u = rand(1000000);
julia> @benchmark conv(u, u)
BenchmarkTools.Trial: 
  memory estimate:  122.08 MiB
  allocs estimate:  156
  --------------
  minimum time:     91.795 ms (2.14% GC)
  median time:      157.906 ms (36.03% GC)
  mean time:        130.520 ms (24.95% GC)
  maximum time:     202.017 ms (55.41% GC)
  --------------
  samples:          39
  evals/sample:     1
```

After changes:
```julia
using DSP, BenchmarkTools
julia> u = rand(1000000);
julia> @benchmark conv(u, u)
BenchmarkTools.Trial: 
  memory estimate:  61.04 MiB
  allocs estimate:  146
  --------------
  minimum time:     76.519 ms (1.51% GC)
  median time:      79.611 ms (1.49% GC)
  mean time:        82.073 ms (3.67% GC)
  maximum time:     191.095 ms (53.96% GC)
  --------------
  samples:          61
  evals/sample:     1
```

While `conv` was already fast, small changes make a big difference when doing Monte Carlo simulations, for example.