@galenlynch 2019-03-17
Also it's amazing how much faster this is than Matlab: for the same problem (convolution of two million element vectors), Matlab takes 51.15 seconds, compared to 0.079 seconds in Julia. Perhaps they're doing time domain convolution or something?