@HDictus 2019-03-26
Similar still creates a new array right? You could consider trying PaddedViews.jl to see if that makes it even better, though it would be an additional dependency of course.
I can't think of a way to improve the readability of it, and I think it's only a small difference in any case.
It seems like it would be a good idea to add some default benchmarks for certain functions, so that one can test the effects of their changes on that without needing to rewrite it. Maybe there's even a way to incorporate that in the CI?