@galenlynch 2019-04-01
@HDictus the benchmarks are a good idea, I think lots of repos have something like that built into CI to detect performance regressions. If you'd want to give it a shot, it would be a welcome addition.

Not sure about the PaddedViews, I imagine you'd have to benchmark them fair carefully to see if slower array access through the view would be worth the memory savings.